+++ Pods/Target Support Files/Pods-BLTNBoard/Pods-BLTNBoard-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Pods_BLTNBoard : NSObject
+@end
+@implementation PodsDummy_Pods_BLTNBoard
+@end