+++ Sources/InterfaceBuilder/BLTNItemAppearance.swift
@@ -77,9 +77,15 @@ import UIKit
     }()
 
     // MARK: - Corner Radius Customization
+    
+    /// The corner radius of the action button set as half height.
+    @objc public var actionButtonCornerRadiusHalfHeigth: Bool = false
 
     /// The corner radius of the action button (default 12).
     @objc public var actionButtonCornerRadius: CGFloat = 12
+    
+    /// The corner radius of the action button set as half height.
+    @objc public var alternativeButtonCornerRadiusHalfHeigth: Bool = false
 
     /// The corner radius of the alternative button (default 12).
     @objc public var alternativeButtonCornerRadius: CGFloat = 12
@@ -88,12 +94,21 @@ import UIKit
 
     /// An optional custom font to use for the title label. Set this to nil to use the system font.
     @objc public var titleFontDescriptor: UIFontDescriptor?
+    
+    /// An optional custom font to use for the title label. Set this to nil to use the system font.
+    @objc public var titleFont: UIFont?
 
     /// An optional custom font to use for the description label. Set this to nil to use the system font.
     @objc public var descriptionFontDescriptor: UIFontDescriptor?
+    
+    /// An optional custom font to use for the description label. Set this to nil to use the system font.
+    @objc public var descriptionFont: UIFont?
 
     /// An optional custom font to use for the buttons. Set this to nil to use the system font.
     @objc public var buttonFontDescriptor: UIFontDescriptor?
+    
+    /// An optional custom font to use for the buttons. Set this to nil to use the system font.
+    @objc public var buttonFont: UIFont?
 
     /**
      * Whether the description text should be displayed with a smaller font.
@@ -132,8 +147,9 @@ extension BLTNItemAppearance {
      */
 
     @objc public func makeTitleFont() -> UIFont {
-
-        if let titleFontDescriptor = self.titleFontDescriptor {
+        if let titleFont = self.titleFont {
+            return titleFont
+        } else if let titleFontDescriptor = self.titleFontDescriptor {
             return UIFont(descriptor: titleFontDescriptor, size: titleFontSize)
         } else {
             return UIFont.systemFont(ofSize: titleFontSize, weight: .medium)
@@ -149,7 +165,9 @@ extension BLTNItemAppearance {
 
         let size = shouldUseCompactDescriptionText ? compactDescriptionFontSize : descriptionFontSize
 
-        if let descriptionFontDescriptor = self.descriptionFontDescriptor {
+        if let descriptionFont = self.descriptionFont {
+            return descriptionFont
+        } else if let descriptionFontDescriptor = self.descriptionFontDescriptor {
             return UIFont(descriptor: descriptionFontDescriptor, size: size)
         } else {
             return UIFont.systemFont(ofSize: size)
@@ -163,7 +181,9 @@ extension BLTNItemAppearance {
 
     @objc public func makeActionButtonFont() -> UIFont {
 
-        if let buttonFontDescriptor = self.buttonFontDescriptor {
+        if let buttonFont = self.buttonFont {
+            return buttonFont
+        } else if let buttonFontDescriptor = self.buttonFontDescriptor {
             return UIFont(descriptor: buttonFontDescriptor, size: actionButtonFontSize)
         } else {
             return UIFont.systemFont(ofSize: actionButtonFontSize, weight: .semibold)
@@ -177,7 +197,9 @@ extension BLTNItemAppearance {
 
     @objc public func makeAlternativeButtonFont() -> UIFont {
 
-        if let buttonFontDescriptor = self.buttonFontDescriptor {
+        if let buttonFont = self.buttonFont {
+            return buttonFont
+        } else if let buttonFontDescriptor = self.buttonFontDescriptor {
             return UIFont(descriptor: buttonFontDescriptor, size: alternativeButtonFontSize)
         } else {
             return UIFont.systemFont(ofSize: alternativeButtonFontSize, weight: .semibold)